chore: add tests for replaceWildcards
